+++ R/convertScene.R
@@ -417,8 +417,6 @@ convertScene <- function(x = scene3d(minimal), width = NULL, height = NULL,
       obj$par3d$viewport$width <- obj$par3d$viewport$width/fullviewport$width
       obj$par3d$viewport$y <- obj$par3d$viewport$y/fullviewport$height
       obj$par3d$viewport$height <- obj$par3d$viewport$height/fullviewport$height
-      if ("user" %in% obj$par3d$mouseMode)
-        warning("User defined mouse callbacks not supported in rglwidget", call.=FALSE)
     }
     if (obj$type == "planes" && nrow(obj$vertices) > 3) {
       obj$vertices <- obj$vertices[1:3,] # These will be redone