+++ src/subscene.cpp
@@ -33,7 +33,7 @@ Subscene::Subscene(Embedding in_viewport, Embedding in_projection, Embedding in_
   modelMatrix.setIdentity();
   projMatrix.setIdentity(); 
   mouseListeners.push_back(this);
-  for (int i=0; i<3; i++) {
+  for (int i=0; i<5; i++) {
     beginCallback[i] = NULL;
     updateCallback[i] = NULL;
     endCallback[i] = NULL;
@@ -48,7 +48,7 @@ Subscene::~Subscene()
 {
   for (std::vector<Subscene*>::iterator i = subscenes.begin(); i != subscenes.end(); ++ i ) 
     delete (*i);
-  for (int i=0; i<3; i++) 
+  for (int i=0; i<5; i++) 
     if (cleanupCallback[i]) 
       (*cleanupCallback[i])(userData + 3*i);
 }
@@ -1076,42 +1076,58 @@ float Subscene::getDistance(const Vertex& v) const
   return (Zrow*vec) / (Wrow*vec);
 }
 
-viewControlPtr Subscene::getButtonBeginFunc(int which) {
-  return getMaster(EM_MOUSEHANDLERS)->ButtonBeginFunc[which];
+viewControlPtr Subscene::getButtonBeginFunc(int button) {
+  return getMaster(EM_MOUSEHANDLERS)->ButtonBeginFunc[button];
 }
 
-void Subscene::buttonBegin(int which, int mouseX, int mouseY)
+void Subscene::buttonBegin(int button, int mouseX, int mouseY)
 {
-  // Rprintf("Subscene %d::buttonBegin %d\n",  getObjID(), which);
-  (this->*getButtonBeginFunc(which))(mouseX, mouseY);
+  (this->*getButtonBeginFunc(button))(mouseX, mouseY);
 }
 
-viewControlPtr Subscene::getButtonUpdateFunc(int which) 
+viewControlPtr Subscene::getButtonUpdateFunc(int button) 
 {  
-  return getMaster(EM_MOUSEHANDLERS)->ButtonUpdateFunc[which];
+  return getMaster(EM_MOUSEHANDLERS)->ButtonUpdateFunc[button];
 }
 
-void Subscene::buttonUpdate(int which, int mouseX, int mouseY)
+void Subscene::buttonUpdate(int button, int mouseX, int mouseY)
 {
-  (this->*getButtonUpdateFunc(which))(mouseX, mouseY);
+  if (button == bnNOBUTTON && needsBegin != mmNONE) {
+    buttonBegin(button, mouseX, mouseY);
+    needsBegin = mmNONE;
+  }
+  (this->*getButtonUpdateFunc(button))(mouseX, mouseY);
 }
 
-viewControlEndPtr Subscene::getButtonEndFunc(int which)
+viewControlEndPtr Subscene::getButtonEndFunc(int button)
 {
-  return getMaster(EM_MOUSEHANDLERS)->ButtonEndFunc[which];  
+  return getMaster(EM_MOUSEHANDLERS)->ButtonEndFunc[button];  
 }
 
-void Subscene::buttonEnd(int which)
+void Subscene::buttonEnd(int button)
 {
-  (this->*getButtonEndFunc(which))();
+  (this->*getButtonEndFunc(button))();
 }
 
 void Subscene::setDefaultMouseMode()
 {
-  setMouseMode(1, mmPOLAR);
-  setMouseMode(2, mmFOV);
-  setMouseMode(3, mmZOOM);
-  setWheelMode(wmNONE);
+  setMouseMode(bnNOBUTTON,mmNONE);
+  setMouseMode(bnLEFT,    mmPOLAR);
+  setMouseMode(bnRIGHT,   mmFOV);
+  setMouseMode(bnMIDDLE,  mmZOOM);
+  setMouseMode(bnWHEEL,   mmNONE);
+
+  needsBegin = mmNONE;
+  busy = false;
+}
+
+bool Subscene::mouseNeedsWatching() {
+  if (mouseMode[bnNOBUTTON] != mmNONE)
+    return true;
+  for (std::vector<Subscene*>::iterator i = subscenes.begin(); i != subscenes.end(); ++ i ) 
+    if ((*i)->mouseNeedsWatching())
+      return true;
+  return false;
 }
 
 void Subscene::setMouseMode(int button, MouseModeID mode)
@@ -1119,54 +1135,67 @@ void Subscene::setMouseMode(int button, MouseModeID mode)
   if (getEmbedding(EM_MOUSEHANDLERS) == EMBED_INHERIT)
     getParent()->setMouseMode(button, mode);
   else {
-    int index = button-1;
-    mouseMode[index] = mode;
+    mouseMode[button] = mode;
+    if (button == bnNOBUTTON)
+      needsBegin = mode;
     switch (mode) {
-    case mmNONE:
-      ButtonBeginFunc[index] = &Subscene::noneBegin;
-      ButtonUpdateFunc[index] = &Subscene::noneUpdate;
-      ButtonEndFunc[index] = &Subscene::noneEnd;
-      break;
-    case mmTRACKBALL:
-      ButtonBeginFunc[index] = &Subscene::trackballBegin;
-      ButtonUpdateFunc[index] = &Subscene::trackballUpdate;
-      ButtonEndFunc[index] = &Subscene::trackballEnd;
-      break;
-    case mmXAXIS:
-    case mmYAXIS:
-    case mmZAXIS:
-      ButtonBeginFunc[index] = &Subscene::oneAxisBegin;
-      ButtonUpdateFunc[index] = &Subscene::oneAxisUpdate;
-      ButtonEndFunc[index] = &Subscene::trackballEnd; // No need for separate function
-      if (mode == mmXAXIS)      axis[index] = Vertex(1,0,0);
-      else if (mode == mmYAXIS) axis[index] = Vertex(0,1,0);
-      else                      axis[index] = Vertex(0,0,1);
-      break;	    	
-    case mmPOLAR:
-      ButtonBeginFunc[index] = &Subscene::polarBegin;
-      ButtonUpdateFunc[index] = &Subscene::polarUpdate;
-      ButtonEndFunc[index] = &Subscene::polarEnd;
-      break;
-    case mmSELECTING:
-      ButtonBeginFunc[index] = &Subscene::mouseSelectionBegin;
-      ButtonUpdateFunc[index] = &Subscene::mouseSelectionUpdate;
-      ButtonEndFunc[index] = &Subscene::mouseSelectionEnd;
-      break;
-    case mmZOOM:
-      ButtonBeginFunc[index] = &Subscene::adjustZoomBegin;
-      ButtonUpdateFunc[index] = &Subscene::adjustZoomUpdate;
-      ButtonEndFunc[index] = &Subscene::adjustZoomEnd;
-      break;
-    case mmFOV:
-      ButtonBeginFunc[index] = &Subscene::adjustFOVBegin;
-      ButtonUpdateFunc[index] = &Subscene::adjustFOVUpdate;
-      ButtonEndFunc[index] = &Subscene::adjustFOVEnd;
-      break;
-    case mmUSER:
-      ButtonBeginFunc[index] = &Subscene::userBegin;
-      ButtonUpdateFunc[index] = &Subscene::userUpdate;
-      ButtonEndFunc[index] = &Subscene::userEnd;
-      break;	    	
+      case mmNONE:
+        ButtonBeginFunc[button] = &Subscene::noneBegin;
+        ButtonUpdateFunc[button] = &Subscene::noneUpdate;
+        ButtonEndFunc[button] = &Subscene::noneEnd;
+        break;
+      case mmTRACKBALL:
+        ButtonBeginFunc[button] = &Subscene::trackballBegin;
+        ButtonUpdateFunc[button] = &Subscene::trackballUpdate;
+        ButtonEndFunc[button] = &Subscene::trackballEnd;
+        break;
+      case mmXAXIS:
+      case mmYAXIS:
+      case mmZAXIS:
+        ButtonBeginFunc[button] = &Subscene::oneAxisBegin;
+        ButtonUpdateFunc[button] = &Subscene::oneAxisUpdate;
+        ButtonEndFunc[button] = &Subscene::trackballEnd; // No need for separate function
+        if (mode == mmXAXIS)      axis[button] = Vertex(1,0,0);
+        else if (mode == mmYAXIS) axis[button] = Vertex(0,1,0);
+        else                      axis[button] = Vertex(0,0,1);
+        break;	    	
+      case mmPOLAR:
+        ButtonBeginFunc[button] = &Subscene::polarBegin;
+        ButtonUpdateFunc[button] = &Subscene::polarUpdate;
+        ButtonEndFunc[button] = &Subscene::polarEnd;
+        break;
+      case mmSELECTING:
+        ButtonBeginFunc[button] = &Subscene::mouseSelectionBegin;
+        ButtonUpdateFunc[button] = &Subscene::mouseSelectionUpdate;
+        ButtonEndFunc[button] = &Subscene::mouseSelectionEnd;
+        break;
+      case mmZOOM:
+        ButtonBeginFunc[button] = &Subscene::adjustZoomBegin;
+        ButtonUpdateFunc[button] = &Subscene::adjustZoomUpdate;
+        ButtonEndFunc[button] = &Subscene::adjustZoomEnd;
+        break;
+      case mmFOV:
+        ButtonBeginFunc[button] = &Subscene::adjustFOVBegin;
+        ButtonUpdateFunc[button] = &Subscene::adjustFOVUpdate;
+        ButtonEndFunc[button] = &Subscene::adjustFOVEnd;
+        break;
+      case mmUSER:
+        ButtonBeginFunc[button] = &Subscene::userBegin;
+        ButtonUpdateFunc[button] = &Subscene::userUpdate;
+        ButtonEndFunc[button] = &Subscene::userEnd;
+        break;
+      case wmPULL:
+        if (button == bnWHEEL)
+          WheelRotateFunc = &Subscene::wheelRotatePull;
+        break;
+      case wmPUSH:
+        if (button == bnWHEEL)
+          WheelRotateFunc = &Subscene::wheelRotatePush;
+        break;
+      case wmUSER2:
+        if (button == bnWHEEL)
+          WheelRotateFunc = &Subscene::userWheel;
+        break;
     }
   }
 }
@@ -1177,16 +1206,15 @@ void Subscene::setMouseCallbacks(int button, userControlPtr begin, userControlPt
   if (getEmbedding(EM_MOUSEHANDLERS) == EMBED_INHERIT)
     getParent()->setMouseCallbacks(button, begin, update, end, cleanup, user);  
   else {
-    int ind = button - 1;
-    if (cleanupCallback[ind])
-      (*cleanupCallback[ind])(userData + 3*ind);
-    beginCallback[ind] = begin;
-    updateCallback[ind] = update;
-    endCallback[ind] = end;
-    cleanupCallback[ind] = cleanup;
-    userData[3*ind + 0] = *(user++);
-    userData[3*ind + 1] = *(user++);
-    userData[3*ind + 2] = *user;
+    if (cleanupCallback[button])
+      (*cleanupCallback[button])(userData + 3*button);
+    beginCallback[button] = begin;
+    updateCallback[button] = update;
+    endCallback[button] = end;
+    cleanupCallback[button] = cleanup;
+    userData[3*button + 0] = *(user++);
+    userData[3*button + 1] = *(user++);
+    userData[3*button + 2] = *user;
     setMouseMode(button, mmUSER);
   }
 }
@@ -1197,48 +1225,19 @@ void Subscene::getMouseCallbacks(int button, userControlPtr *begin, userControlP
   if (getEmbedding(EM_MOUSEHANDLERS) == EMBED_INHERIT)
     getParent()->getMouseCallbacks(button, begin, update, end, cleanup, user); 
   else {
-    int ind = button - 1;
-    *begin = beginCallback[ind];
-    *update = updateCallback[ind];
-    *end = endCallback[ind];
-    *cleanup = cleanupCallback[ind];
-    *(user++) = userData[3*ind + 0];
-    *(user++) = userData[3*ind + 1];
-    *(user++) = userData[3*ind + 2];
+    *begin = beginCallback[button];
+    *update = updateCallback[button];
+    *end = endCallback[button];
+    *cleanup = cleanupCallback[button];
+    *(user++) = userData[3*button + 0];
+    *(user++) = userData[3*button + 1];
+    *(user++) = userData[3*button + 2];
   }
 } 
 
 MouseModeID Subscene::getMouseMode(int button)
 {
-  return getMaster(EM_MOUSEHANDLERS)->mouseMode[button-1];
-}
-
-void Subscene::setWheelMode(WheelModeID mode)
-{
-  if (getEmbedding(EM_MOUSEHANDLERS) == EMBED_INHERIT)
-    getParent()->setWheelMode(mode); 
-  else {
-    wheelMode = mode;
-    switch (mode) {
-    case wmNONE:
-      WheelRotateFunc = &Subscene::wheelRotateNone;
-      break;
-    case wmPULL:
-      WheelRotateFunc = &Subscene::wheelRotatePull;
-      break;
-    case wmPUSH:
-      WheelRotateFunc = &Subscene::wheelRotatePush;
-      break;
-    case wmUSER:
-      WheelRotateFunc = &Subscene::userWheel;
-      break;
-    }
-  }
-}
-
-WheelModeID Subscene::getWheelMode()
-{
-  return getMaster(EM_MOUSEHANDLERS)->wheelMode;
+  return getMaster(EM_MOUSEHANDLERS)->mouseMode[button];
 }
 
 void Subscene::setWheelCallback(userWheelPtr wheel, void* user)
@@ -1248,7 +1247,7 @@ void Subscene::setWheelCallback(userWheelPtr wheel, void* user)
   else {
     wheelCallback = wheel;
     wheelData = user;
-    setWheelMode(wmUSER); 
+    setMouseMode(bnWHEEL, wmUSER2); 
   }
 }
 
@@ -1347,7 +1346,6 @@ void Subscene::trackballBegin(int mouseX, int mouseY)
 void Subscene::trackballUpdate(int mouseX, int mouseY)
 {
   rotCurrent = screenToVector(pviewport.width,pviewport.height,mouseX,mouseY);
-
   for (unsigned int i = 0; i < mouseListeners.size(); i++) {
     Subscene* sub = mouseListeners[i];
     if (sub) {
@@ -1376,7 +1374,6 @@ void Subscene::oneAxisBegin(int mouseX, int mouseY)
 void Subscene::oneAxisUpdate(int mouseX, int mouseY)
 {
   rotCurrent = screenToVector(pviewport.width,pviewport.height,mouseX,pviewport.height/2);
-  
   for (unsigned int i = 0; i < mouseListeners.size(); i++) {
     Subscene* sub = mouseListeners[i];
     if (sub) {
@@ -1492,20 +1489,26 @@ void Subscene::wheelRotatePush(int dir)
 
 void Subscene::wheelRotate(int dir)
 {
-  (this->*WheelRotateFunc)(dir);
+  int mode = getMouseMode(bnWHEEL);
+  if (mode >= wmPULL)
+    (this->*WheelRotateFunc)(dir);
+  else {
+    // Need to fake a click and release dir = 1 rotates away, dir = 2 rotates towards
+    buttonBegin(bnWHEEL, pviewport.width / 2, pviewport.height / 2);
+    buttonUpdate(bnWHEEL, pviewport.width / 2, pviewport.height / 2 + 10*(dir == 1 ? 1 : -1));
+    buttonEnd(bnWHEEL);
+  }
 }
 
 void Subscene::userBegin(int mouseX, int mouseY)
 {
-  int ind = drag - 1;
   Subscene* master = getMaster(EM_MOUSEHANDLERS);
-  beginCallback[ind] = master->beginCallback[ind];
-  void* this_userData = master->userData[3*ind+0];
-  // Rprintf("userBegin in %d with ind=%d\n", getObjID(), ind);
+  beginCallback[drag] = master->beginCallback[drag];
+  void* this_userData = master->userData[3*drag+0];
   activeButton = drag;
-  if (beginCallback[ind]) {
+  if (beginCallback[drag]) {
     busy = true;
-    (*beginCallback[ind])(this_userData, mouseX, pviewport.height-mouseY);
+    (*beginCallback[drag])(this_userData, mouseX, pviewport.height-mouseY);
     busy = false;
   }
 }
@@ -1513,25 +1516,23 @@ void Subscene::userBegin(int mouseX, int mouseY)
 
 void Subscene::userUpdate(int mouseX, int mouseY)
 {
-  int ind = activeButton - 1;
   Subscene* master = getMaster(EM_MOUSEHANDLERS);
-  updateCallback[ind] = master->updateCallback[ind];
-  void* this_userData = master->userData[3*ind+1];
-  if (!busy && updateCallback[ind]) {
+  updateCallback[activeButton] = master->updateCallback[activeButton];
+  void* this_userData = master->userData[3*activeButton+1];
+  if (!busy && updateCallback[activeButton]) {
     busy = true;
-    (*updateCallback[ind])(this_userData, mouseX, pviewport.height-mouseY);
+    (*updateCallback[activeButton])(this_userData, mouseX, pviewport.height-mouseY);
     busy = false;
   }
 }
 
 void Subscene::userEnd()
 {
-  int ind = activeButton - 1;
   Subscene* master = getMaster(EM_MOUSEHANDLERS);
-  endCallback[ind] = master->endCallback[ind];
-  void* this_userData = master->userData[3*ind+2];
-  if (endCallback[ind])
-    (*endCallback[ind])(this_userData);
+  endCallback[activeButton] = master->endCallback[activeButton];
+  void* this_userData = master->userData[3*activeButton+2];
+  if (endCallback[activeButton])
+    (*endCallback[activeButton])(this_userData);
 }
 
 void Subscene::userWheel(int dir)
@@ -1550,12 +1551,9 @@ void Subscene::adjustZoomBegin(int mouseX, int mouseY)
 void Subscene::adjustZoomUpdate(int mouseX, int mouseY)
 {
   int dy = mouseY - zoomBaseY;
-  // Rprintf("adjustZoomUpdate by %d for %d\n", dy, getObjID());
   for (unsigned int i = 0; i < mouseListeners.size(); i++) {
-    // Rprintf("adjustZoomUpdate: mouseListeners[%d]=%d\n", i, mouseListeners[i]);
     Subscene* sub = mouseListeners[i];
     if (sub) {
-//      Rprintf("found it\n");
       UserViewpoint* this_userviewpoint = sub->getUserViewpoint();
       
       float zoom = clamp ( this_userviewpoint->getZoom() * exp(dy*ZOOM_PIXELLOGSTEP), ZOOM_MIN, ZOOM_MAX);
@@ -1609,4 +1607,9 @@ void Subscene::mouseSelectionEnd()
   if (selectState == msABORT) return;
   
   selectState = msDONE;
+}
+
+Subscene* Subscene::getRootSubscene()
+{
+  return parent ? parent->getRootSubscene() : this;
 }